% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ch}
\alias{ch}
\title{abbreviations: ds - dataset, patt - pattern, pattNeg - negative pattern,  #ex: ,  #e  - example ,  #en  - example don't run}
\usage{
ch()
}
\description{
abbreviations: ds - dataset, patt - pattern, pattNeg - negative pattern,  #ex: ,  #e  - example ,  #en  - example don't run
}

