\name{HaLaP-package}
\alias{HaLaP-package}
\alias{HaLaP}
\docType{package}
\title{ HaLaP 
}
\description{
 HaLaP - Miscellaneous aliases and functions that we use the most
    often during project initiation, in progress, leaving and coming back for a
    book "Handling Large R Projects"
}
\details{
\tabular{ll}{
Package: \tab HaLaP\cr
Type: \tab Package\cr
Version: \tab 0.1.2\cr
Date: \tab 2014-08-17\cr
License: \tab  GPL-2\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{Alex Zolotovitski <alex.zolot@gmail.com>

Maintainer: Alex Zolotovitski <alex.zolot@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}

