% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sa2}
\alias{sa2}
\alias{saaa}
\title{sa2= saaa}
\usage{
sa2(...)
}
\arguments{
\item{...}{0, 1, or 2 args
0 - save.image
1 - if char with  le=1 it is file , else object to save
2 - first is obj, second is file name}
}
\description{
wrapper for save or save.image
}

