% Generated by roxygen2 (4.0.1): do not edit by hand
\name{srt}
\alias{sort.data.frame}
\alias{sortt}
\alias{srt}
\title{srt= sortt= sort.data.frame}
\usage{
srt(x, by)
}
\description{
Sort Data Frame
Author: Kevin Wright
with some ideas from Andy Liaw
http://tolstoy.newcastle.edu.au/R/help/04/07/1076.html
    Sorting is left to right in the formula
Useage is:
}

