\docType{package}
\name{HLP}
\alias{HLP}
\alias{HLP-package}
\title{HLP  - Miscellaneous aliases and functions}
\description{
HLP - Miscellaneous aliases and functions that I use the
most often during project initiation, in progress, leaving
and coming back for a book "Handling Large R Projects"
}
\author{
Alex Zolotovitski <alex@zolot.us>
}
\keyword{aliases}

